src/requirements_py312.txt
# ============================================================================
# DEPENDENCIAS — Fase 2 NSD + Stable Diffusion 2.1 unCLIP
# ============================================================================
# Instalación:
#   py -3.12 -m pip install -r requirements_py312.txt
#
# GPU (CUDA 12.1):
#   py -3.12 -m pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# ============================================================================

--extra-index-url https://download.pytorch.org/whl/cu118

# ----------------------------------------------------------------------------
# Core deep learning
# ----------------------------------------------------------------------------
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
einops>=0.7.0

# ----------------------------------------------------------------------------
# Stable Diffusion 2.1 unCLIP stack
# ----------------------------------------------------------------------------
diffusers>=0.27.0
accelerate>=0.28.0
transformers>=4.38.0
safetensors>=0.4.0

# ----------------------------------------------------------------------------
# fMRI / NSD
# ----------------------------------------------------------------------------
nibabel>=5.0
nilearn>=0.10
h5py>=3.0
scikit-learn>=1.3

# ----------------------------------------------------------------------------
# Datos e imágenes
# ----------------------------------------------------------------------------
Pillow>=10.0
scipy>=1.11
scikit-image>=0.22
lpips>=0.1.4
ftfy>=6.0
regex>=2023.0
tqdm>=4.66
requests>=2.31

# ----------------------------------------------------------------------------
# Visualización
# ----------------------------------------------------------------------------
matplotlib>=3.7
pandas>=2.0
seaborn>=0.13

# ----------------------------------------------------------------------------
# Logging
# ----------------------------------------------------------------------------
coloredlogs>=15.0
rich>=13.0
